Move `StAnnTxCache` into `EraTx` and:
* Change `getCacheStAnnTx` to `cacheStAnnTxG` as well for consistency
  and move it to `EraTx`
* add cache to `StAnnTx SubTx`
