+++ eras/babbage/impl/src/Cardano/Ledger/Babbage/Rules/Ledgers.hs
@@ -134,7 +134,7 @@ ledgersTransition = do
           newLedgerState <-
             trans @(EraRule "LEDGER" era) $
               TRC (Shelley.LedgerEnv slot (Just epochNo) ix pp account, curLedgerState, stAnnTx)
-          pure (newLedgerState, getCacheStAnnTx stAnnTx)
+          pure (newLedgerState, stAnnTx ^. cacheStAnnTxG)
       )
       (initLedgerState, mempty)
       $ zip [minBound ..]
