+++ eras/conway/impl/src/Cardano/Ledger/Conway/UTxO.hs
@@ -17,8 +17,6 @@ module Cardano.Ledger.Conway.UTxO (
   getConwayMinFeeTxUtxo,
 ) where
 
-import Cardano.Ledger.Alonzo.Plutus.Context (SupportedPlutusRunnable)
-import Cardano.Ledger.Alonzo.Tx (AlonzoStAnnTx (..))
 import Cardano.Ledger.Alonzo.UTxO (
   AlonzoEraUTxO (..),
   AlonzoScriptsNeeded (..),
@@ -120,7 +118,6 @@ conwayProducedValue pp isStakePool txBody =
 
 instance EraUTxO ConwayEra where
   type ScriptsNeeded ConwayEra = AlonzoScriptsNeeded ConwayEra
-  type StAnnTxCache ConwayEra = Map.Map ScriptHash (SupportedPlutusRunnable ConwayEra)
 
   getConsumedValue = getConsumedMaryValue
 
@@ -135,8 +132,6 @@ instance EraUTxO ConwayEra where
   getWitsVKeyNeeded _ = getConwayWitsVKeyNeeded
 
   getMinFeeTxUtxo = getConwayMinFeeTxUtxo
-
-  getCacheStAnnTx = asatPlutusRunnableCache
 
 instance AlonzoEraUTxO ConwayEra where
   getSupplementalDataHashes = getBabbageSupplementalDataHashes
